[BE] refactor: ProjectLabel과 ModelLabel을 사용 주체에 맞게 분리
